% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CellRanger-package}
\alias{CellRanger}
\alias{CellRanger-package}
\title{CellRanger}
\description{
Toolkit for 10X Genomics single-cell RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://steinbaugh.com/CellRanger/}
  \item \url{https://github.com/steinbaugh/CellRanger/}
  \item Report bugs at \url{https://github.com/steinbaugh/CellRanger/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
\keyword{internal}
