% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanger.R
\name{CellRanger}
\alias{CellRanger}
\title{10X Genomics Cell Ranger Data Set}
\usage{
CellRanger(uploadDir, format = c("mtx", "hdf5"), filtered = TRUE,
  sampleMetadataFile = NULL, organism = NULL, ensemblRelease = NULL,
  genomeBuild = NULL, refdataDir = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL,
  interestingGroups = "sampleName")
}
\arguments{
\item{uploadDir}{\code{string}. Path to Cell Ranger output directory
(final upload). This directory path must contain
\code{filtered_gene_bc_matrices*} as a child directory.}

\item{format}{\code{string}. Output format, either MatrixMarket ("\code{mtx}") or HDF5
("\code{hdf5}").}

\item{filtered}{\code{boolean}. Use filtered (recommended) or raw counts. Note
that raw counts still contain only whitelisted cellular barcodes.}

\item{organism}{\code{string}. Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{ensemblRelease}{\code{scalar integer}. Ensembl release version (e.g. \code{90}).}

\item{refdataDir}{\code{string} or \code{NULL}. Directory path to Cell Ranger reference
annotation data.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}
}
\value{
\code{CellRanger}.
}
\description{
Extends \code{SingleCellExperiment}, with additional validity checks on the
\code{metadata()} slot.
}
\details{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Cell Ranger}
counts into a \code{SingleCellExperiment} object.
}
\section{Directory structure}{

Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for all datasets, even those that only
contain a single sample:

\preformatted{
file.path(
    "<uploadDir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "<genomeBuild>",
    "matrix.mtx"
)
}
}

\section{Sample metadata}{

A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{

We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{\link[=rowRanges]{rowRanges()}}. Otherwise, the function will attempt to use the
most current annotations available from Ensembl, and some gene IDs may not
match, due to deprecation in the current Ensembl release.
}

\examples{
uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
x <- CellRanger(uploadDir)
print(x)
}
\seealso{
\code{\link[=CellRanger]{CellRanger()}}.
}
\author{
Michael Steinbaugh
}
