% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellRangerATAC-class}
\alias{CellRangerATAC-class}
\title{Cell Ranger ATAC-seq data set}
\description{
Contains single-cell ATAC-seq data.
}
\examples{
showClass("CellRangerATAC")
}
\author{
Michael Steinbaugh
}
