% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/CellRangerRNASeq.R
\docType{class}
\name{CellRangerRNASeq-class}
\alias{CellRangerRNASeq-class}
\alias{CellRangerRNASeq}
\title{Cell Ranger RNA-seq data set}
\usage{
CellRangerRNASeq(dir, filtered = TRUE, sampleMetadataFile = NULL,
  organism = NULL, ensemblRelease = NULL, genomeBuild = NULL,
  refdataDir = NULL, gffFile = NULL, transgeneNames = NULL,
  spikeNames = NULL, interestingGroups = "sampleName")
}
\arguments{
\item{dir}{\code{character(1)}.
Path to Cell Ranger output directory (final upload). This directory path
must contain \code{filtered_gene_bc_matrices*} as a child directory.}

\item{filtered}{\code{logical(1)}.
Use filtered (recommended) or raw counts. Note that raw counts still
contain only whitelisted cellular barcodes.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available. Note: don't pass in UCSC build
IDs (e.g. \code{"hg38"}).}

\item{refdataDir}{\code{character(1)} or \code{NULL}.
Directory path to Cell Ranger reference annotation data.}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF
(GFFv2) instead of a GFFv3 file if possible.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{CellRanger}.
}
\description{
Contains UMI droplet-based single-cell RNA-seq data.
}
\details{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Cell Ranger} output
for a Chromium data set into a \code{SingleCellExperiment} object.
}
\note{
Updated 2019-08-01.
}
\section{Directory structure for multiple samples}{


Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for datasets containing multiple
samples that have not been aggregated into a single matrix with \code{aggr}.

Cell Ranger v3 output:

\preformatted{
| <dir>/
|-- <sampleName>/
|---- outs/
|------ filtered_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ filtered_feature_bc_matrix.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ raw_feature_bc_matrix.h5
|------ web_summary.html
)
}

Cell Ranger v2 output:

\preformatted{
| <dir>/
|-- <sampleName>/
|---- outs/
|------ filtered_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ filtered_gene_bc_matrices_h5.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ raw_gene_bc_matrices_h5.h5
)
}
}

\section{Sample metadata}{


A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{


We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{rowRanges}. Otherwise, the function will attempt to use the
most current annotations available from Ensembl, and some gene IDs may not
match, due to deprecation in the current Ensembl release.
}

\examples{
showClass("CellRangerRNASeq")
dir <- system.file("extdata/cellranger", package = "Chromium")
x <- CellRangerRNASeq(dir)
print(x)
}
\author{
Michael Steinbaugh

Michael Steinbaugh
}
