% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMPLE.R
\name{CellRangerRNASeqSimple}
\alias{CellRangerRNASeqSimple}
\title{Import Cell Ranger simple mode}
\usage{
CellRangerRNASeqSimple(matrixFile)
}
\arguments{
\item{matrixFile}{\code{character(1)}.
File path to:
\itemize{
\item HDF5 matrix file (e.g. \code{filtered_feature_bc_matrix.h5})
\item MTX sparse matrix file (e.g. \code{matrix.mtx.gz}).
}

Filtered (\strong{recommended}) or unfiltered matrix is supported.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Import Cell Ranger simple mode
}
