% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Chromium-class}
\alias{Chromium-class}
\title{10X Genomics Chromium Single-Cell RNA-Seq Data Set}
\description{
Contains single-cell RNA-seq counts processed by the Cell Ranger pipeline.
}
\details{
Extends \code{SingleCellExperiment}, with additional validity checks on the
\code{metadata()} slot.
}
\seealso{
\code{Chromium()}.
}
\author{
Michael Steinbaugh
}
\concept{S4 classes}
