% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Chromium-package}
\title{Chromium}
\description{
Toolkit for 10X Genomics single-cell RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chromium.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/Chromium/}
  \item Report bugs at \url{https://github.com/acidgenomics/Chromium/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
