% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromium.R
\name{Chromium}
\alias{Chromium}
\title{10X Genomics Chromium Single-Cell RNA-Seq Data Set}
\usage{
Chromium(dir, format = c("mtx", "hdf5"), filtered = TRUE,
  sampleMetadataFile = NULL, organism = NULL, ensemblRelease = NULL,
  genomeBuild = NULL, refdataDir = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL,
  interestingGroups = "sampleName")
}
\arguments{
\item{dir}{\code{string}. Path to Cell Ranger output directory
(final upload). This directory path must contain
\code{filtered_gene_bc_matrices*} as a child directory.}

\item{format}{\code{string}. Output format, either MatrixMarket ("\code{mtx}") or HDF5
("\code{hdf5}").}

\item{filtered}{\code{boolean}. Use filtered (recommended) or raw counts. Note
that raw counts still contain only whitelisted cellular barcodes.}

\item{sampleMetadataFile}{\code{string}. Sample metadata file path. CSV
or TSV is preferred, but Excel worksheets are also supported. Check the
documentation for conventions and required columns.}

\item{organism}{\code{string}. Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{ensemblRelease}{\code{scalar integer}. Ensembl release version (e.g. \code{90}).
We recommend setting this value if possible, for improved reproducibility.
When left unset, the latest release available via AnnotationHub/ensembldb
is used. Note that the latest version available can vary, depending on the
versions of AnnotationHub and ensembldb in use.}

\item{genomeBuild}{\code{string}. Ensembl genome build assembly nam (e.g.
\code{"GRCh38"}). If set \code{NULL}, defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{refdataDir}{\code{string} or \code{NULL}. Directory path to Cell Ranger reference
annotation data.}

\item{gffFile}{\code{string}. GFF/GTF (General Feature Format) file.
Generally, we recommend using a GTF (GFFv2) instead of a GFFv3 file if
possible.}

\item{transgeneNames}{\code{character}. Vector indicating which assay
rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}. Vector indicating which assay rows denote
spike-in sequences (e.g. ERCCs).}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}
}
\value{
\code{Chromium}.
}
\description{
Contains single-cell RNA-seq counts processed by the Cell Ranger pipeline.
}
\details{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Cell Ranger}
counts into a \code{SingleCellExperiment} object.
}
\section{Directory structure}{

Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for all datasets, even those that only
contain a single sample:

\preformatted{
file.path(
    "<dir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "<genomeBuild>",
    "matrix.mtx"
)
}
}

\section{Sample metadata}{

A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{

We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{rowRanges()}. Otherwise, the function will attempt to use the
most current annotations available from Ensembl, and some gene IDs may not
match, due to deprecation in the current Ensembl release.
}

\examples{
dir <- system.file("extdata/cellranger", package = "Chromium")
x <- Chromium(dir)
print(x)
}
\seealso{
\code{Chromium()}.
}
\author{
Michael Steinbaugh
}
