% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{ChromiumTestsURL}
\alias{ChromiumTestsURL}
\title{Chromium test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
ChromiumTestsURL
}
\description{
Chromium test data URL
}
\examples{
ChromiumTestsURL
}
\keyword{internal}
