% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr}
\alias{cr}
\title{10X Genomics Cell Ranger Example Data Set}
\format{An object of class \code{numeric} of length 1.}
\usage{
cr
}
\description{
10X Genomics Cell Ranger Example Data Set
}
\examples{
data(cr)
cr
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
