% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract}
\alias{extract}
\alias{[,CellRanger,ANY,ANY,ANY-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{CellRanger,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{Object.}

\item{i, j, ...}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer or whole numbers as
    by \code{\link[base]{as.integer}} or for large values by \code{\link[base]{trunc}}
    (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{relevant for matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\value{
\code{CellRanger}.
}
\description{
Extract genes by row and cells by column.
}
\details{
Refer to \code{cellToSample()} and \code{selectSamples()} if sample-level extraction is
desired. Note that \code{sampleId} is slotted into \code{colData()} and defines the
cell-to-sample mappings.

Unfiltered cellular barcode distributions for the entire dataset, including
cells not kept in the matrix will be dropped in favor of the \code{nCount} column
of \code{colData}.
}
\note{
Updated 2023-09-25.
}
\examples{
data(pbmc_v3)

## CellRanger ====
object <- pbmc_v3

cells <- head(colnames(object), 100L)
head(cells)
genes <- head(rownames(object), 100L)
head(genes)

## Subset by cell identifiers.
object[, cells]

## Subset by genes.
object[genes, ]

## Subset by both genes and cells.
object[genes, cells]
}
\references{
{\if{html}{\out{<span id="reference+Extract.Rd+R+3ABecker+2BChambers+2BWilks+3A1988"></span>}}Becker RA, Chambers JM, Wilks AR (1988).
\emph{The New S Language}.
Chapman and Hall/CRC, London.
ISBN 053409192X.}
}
\author{
Michael Steinbaugh
}
