% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCountsHDF5.R
\name{importCountsHDF5}
\alias{importCountsHDF5}
\title{Import Cell Ranger count matrix from HDF5 file}
\usage{
importCountsHDF5(file)
}
\value{
\code{sparseMatrix}.
Cell barcodes in the columns, features (i.e. genes) in the rows.
}
\description{
Import Cell Ranger count matrix from HDF5 file
}
\note{
Updated 2019-07-31.
}
\examples{
## > x <- importCountsHDF5(file = "filtered_feature_bc_matrix.h5")
## > dim(x)
}
\seealso{
\code{cellrangerRkit::get_matrix_from_h5()}
}
