% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCountsMTX.R
\name{importCountsMTX}
\alias{importCountsMTX}
\title{Import Cell Ranger count matrix from MTX file}
\usage{
importCountsMTX(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\description{
Import Cell Ranger count matrix from MTX file
}
\note{
Data import using HDF5 file is now recommended over this approach.

Updated 2019-08-01.
}
\section{Matrix Market Exchange (MEX/MTX) format}{


Loading from this matrix requires sidecar files containing cell barcodes and
feature (i.e. gene) identifiers.

Cell Ranger v3:
\itemize{
\item \code{barcodes.tsv.gz}: Cell barcodes.
\item \code{features.tsv.gz}: Feature identifiers.
}

Cell Ranger v2:
\itemize{
\item \code{barcodes.tsv}: Cell barcodes.
\item \code{genes.tsv}: Gene identifiers.
}
}

\examples{
## > x <- importCountsMTX(file = "matrix.mtx.gz")
## > dim(x)
}
