% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_v2}
\alias{pbmc4k_v2}
\title{10X Genomics Chromium example data set}
\format{An object of class \code{CellRanger} with 500 rows and 100 columns.}
\usage{
pbmc4k_v2
}
\description{
Peripheral blood mononuclear cells (PBMCs).
}
\examples{
data(pbmc4k_v2)
pbmc4k_v2
}
\keyword{datasets}
