% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_v3}
\alias{pbmc_v3}
\title{10X Genomics Chromium example data set}
\format{An object of class \code{CellRanger} with 500 rows and 100 columns.}
\usage{
pbmc_v3
}
\description{
Peripheral blood mononuclear cells (PBMCs).
}
\details{
Chromium v3 chemistry.
}
\examples{
data(pbmc_v3)
pbmc_v3
}
\keyword{datasets}
