% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellRanger-class}
\alias{CellRanger-class}
\title{Cell Ranger RNA-seq data set}
\description{
Contains UMI droplet-based single-cell RNA-seq data.
}
\examples{
showClass("CellRanger")
}
\author{
Michael Steinbaugh
}
