% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellRanger}
\alias{CellRanger}
\title{Cell Ranger RNA-seq data set}
\usage{
CellRanger(
  dir,
  filtered = TRUE,
  organism = NULL,
  ensemblRelease = NULL,
  genomeBuild = NULL,
  gffFile = NULL,
  refdataDir = NULL,
  samples = NULL,
  censorSamples = NULL,
  sampleMetadataFile = NULL,
  transgeneNames = NULL,
  interestingGroups = "sampleName"
)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory path to Cell Ranger output.}

\item{filtered}{\code{logical(1)}.
Use filtered (recommended) or raw counts. Note that raw counts still
contain only whitelisted cellular barcodes.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file.
Generally, we recommend using GTF (GFFv2) instead of GFFv3.}

\item{refdataDir}{\code{character(1)} or \code{NULL}.
Directory path to Cell Ranger reference annotation data.}

\item{samples}{\code{character}.
Sample identifiers.}

\item{censorSamples}{\code{character}.
Specify a subset of samples to censor.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}
}
\value{
\code{CellRanger}.
}
\description{
Contains UMI droplet-based single-cell RNA-seq data.
}
\details{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Cell Ranger} output
for a Chromium data set into a \code{SingleCellExperiment} object.

Currently supports loading of a single genome.
}
\note{
Updated 2022-06-07.
}
\section{Directory structure for multiple samples}{


Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for datasets containing multiple
samples that have not been aggregated into a single matrix with \code{aggr}.

Cell Ranger v3 output:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| <dir>/
|-- <sampleName>/
|---- SC_RNA_COUNTER_CS/
|---- outs/
|------ filtered_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ filtered_feature_bc_matrix.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ raw_feature_bc_matrix.h5
|------ web_summary.html
}\if{html}{\out{</div>}}

Cell Ranger v2 output:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| <dir>/
|-- <sampleName>/
|---- SC_RNA_COUNTER_CS/
|---- outs/
|------ filtered_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ filtered_gene_bc_matrices_h5.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ raw_gene_bc_matrices_h5.h5
}\if{html}{\out{</div>}}
}

\section{Sample metadata}{


A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{


We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges()}}. Otherwise, the
function will attempt to use the most current annotations available from
Ensembl, and some gene IDs may not match, due to deprecation in the current
Ensembl release.
}

\examples{
dir <- system.file("extdata", "cellranger_v3", package = "Chromium")
x <- CellRanger(dir)
print(x)
}
\seealso{
\itemize{
\item https://support.10xgenomics.com/single-cell-gene-expression/
}
}
