% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellRanger}
\alias{CellRanger}
\title{Cell Ranger RNA-seq data set}
\usage{
CellRanger(dir, filtered = TRUE, organism = NULL,
  ensemblRelease = NULL, genomeBuild = NULL, gffFile = NULL,
  refdataDir = NULL, samples = NULL, censorSamples = NULL,
  sampleMetadataFile = NULL, transgeneNames = NULL,
  spikeNames = NULL, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{dir}{\code{character(1)}.
Path to Cell Ranger output directory (final upload). This directory path
must contain \code{filtered_gene_bc_matrices*} as a child directory.}

\item{filtered}{\code{logical(1)}.
Use filtered (recommended) or raw counts. Note that raw counts still
contain only whitelisted cellular barcodes.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. "\code{Homo sapiens}").}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{gffFile}{\code{character(1)}.
GFF/GTF (General Feature Format) file.
Generally, we recommend using GTF (GFFv2) instead of GFFv3.}

\item{refdataDir}{\code{character(1)} or \code{NULL}.
Directory path to Cell Ranger reference annotation data.}

\item{samples}{\code{character}.
Sample identifiers.}

\item{censorSamples}{\code{character}.
Specify a subset of samples to censor.}

\item{sampleMetadataFile}{\code{character(1)}.
Sample metadata file path. CSV or TSV is preferred, but Excel worksheets
are also supported. Check the documentation for conventions and required
columns.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{CellRanger}.
}
\description{
Contains UMI droplet-based single-cell RNA-seq data.
}
\details{
Read \href{https://www.10xgenomics.com/software/}{10x Genomics Cell Ranger} output
for a Chromium data set into a \code{SingleCellExperiment} object.
}
\note{
Updated 2019-08-07.

Updated 2019-08-07.
}
\section{Directory structure for multiple samples}{


Cell Ranger can vary in its output directory structure, but we're requiring a
single, consistent directory structure for datasets containing multiple
samples that have not been aggregated into a single matrix with \code{aggr}.

Cell Ranger v3 output:

\preformatted{
| <dir>/
|-- <sampleName>/
|---- outs/
|------ filtered_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ filtered_feature_bc_matrix.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_feature_bc_matrix/
|-------- barcodes.tsv.gz
|-------- features.tsv.gz
|-------- matrix.mtx.gz
|------ raw_feature_bc_matrix.h5
|------ web_summary.html
)
}

Cell Ranger v2 output:

\preformatted{
| <dir>/
|-- <sampleName>/
|---- outs/
|------ filtered_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ filtered_gene_bc_matrices_h5.h5
|------ metrics_summary.csv
|------ molecule_info.h5
|------ possorted_genome_bam.bam
|------ possorted_genome_bam.bam.bai
|------ raw_gene_bc_matrices/
|-------- <genomeBuild>/
|---------- barcodes.tsv
|---------- genes.tsv
|---------- matrix.mtx
|------ raw_gene_bc_matrices_h5.h5
)
}
}

\section{Sample metadata}{


A user-supplied sample metadata file defined by \code{sampleMetadataFile} is
required for multiplexed datasets. Otherwise this can be left \code{NULL}, and
minimal sample data will be used, based on the directory names.
}

\section{Reference data}{


We strongly recommend supplying the corresponding reference data required for
Cell Ranger with the \code{refdataDir} argument. It will convert the gene
annotations defined in the GTF file into a \code{GRanges} object, which get
slotted in \code{\link[SummarizedExperiment:rowRanges]{rowRanges()}}. Otherwise, the
function will attempt to use the most current annotations available from
Ensembl, and some gene IDs may not match, due to deprecation in the current
Ensembl release.
}

\examples{
dir <- system.file("extdata/cellranger", package = "Chromium")
x <- CellRanger(dir)
print(x)
}
\author{
Michael Steinbaugh
}
