% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Chromium-class}
\alias{Chromium-class}
\title{10X Genomics Chromium}
\description{
Extends \code{SingleCellExperiment}.
}
\examples{
showClass("Chromium")
}
\author{
Michael Steinbaugh
}
