% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract}
\alias{extract}
\alias{[,CellRanger,ANY,ANY,ANY-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{CellRanger,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{...}{Additional arguments.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\value{
\code{CellRanger}.
}
\description{
Extract genes by row and cells by column.
}
\details{
Refer to \code{\link[basejump:reexports]{cell2sample()}} and
\code{\link[basejump:reexports]{selectSamples()}} if sample-level extraction is
desired. Note that \code{sampleId} is slotted into
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} and defines the cell-to-sample
mappings.

Unfiltered cellular barcode distributions for the entire dataset, including
cells not kept in the matrix will be dropped in favor of the \code{nCount} column
of \code{colData}.
}
\note{
Updated 2022-06-07.
}
\examples{
data(pbmc_v3)

## CellRanger ====
object <- pbmc_v3

cells <- head(colnames(object), 100L)
head(cells)
genes <- head(rownames(object), 100L)
head(genes)

## Subset by cell identifiers.
object[, cells]

## Subset by genes.
object[genes, ]

## Subset by both genes and cells.
object[genes, cells]
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
\author{
Michael Steinbaugh
}
