% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{10X Genomics Chromium Example Data Set
Peripheral blood mononuclear cells (PBMCs).}
\format{An object of class \code{Chromium} with 500 rows and 100 columns.}
\usage{
pbmc
}
\description{
10X Genomics Chromium Example Data Set
Peripheral blood mononuclear cells (PBMCs).
}
\examples{
data(pbmc)
pbmc
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
